doing some cleanup
conflict removal
